Added edit contact feature

+++ app/contact/routes.py
@@ -1,12 +1,12 @@
 from app.contact import bp
 from flask_login import login_required
-from flask import render_template, flash, redirect, url_for, request
+from flask import render_template, flash, redirect, url_for, request, abort
 from app.models.contact import Contact
-from sqlalchemy import select
+from sqlalchemy import select, update
 from flask_login import current_user
 from app import db
 from icecream import ic
-from app.contact.forms import NewContactForm
+from app.contact.forms import NewContactForm, UpdateContactForm
 
 @bp.route('/index', methods=['GET', 'POST'])
 @login_required
@@ -51,13 +51,43 @@ def index():
 @bp.route('/contact/delete/<int:item_id>', methods=['GET', 'POST'])
 @login_required
 def delete(item_id):
-    try:
-        stmt = select(Contact).where(Contact.id==item_id)
-        contact_object_to_delete = db.session.scalar(stmt)
-        db.session.delete(contact_object_to_delete)
-        db.session.commit()
-    except Exception as delete_error:
-        ic(delete_error)
-        flash("Error in deleting item.")
-        return redirect(url_for('contact.index'))
+    if current_user.is_authenticated:
+        try:
+            stmt = select(Contact).where(Contact.id==item_id)
+            contact_object_to_delete = db.session.scalar(stmt)
+            db.session.delete(contact_object_to_delete)
+            db.session.commit()
+            flash("Item deleted successfully.")
+        except Exception as delete_error:
+            ic(delete_error)
+            flash("Error in deleting item.")
+            return redirect(url_for('contact.index'))
+    else:
+        abort(401)
     return redirect(url_for('contact.index'))
+
+@bp.route('/contact/update/<int:item_id>', methods=['GET', 'POST'])
+@login_required
+def update(item_id):
+    contact_item = Contact.query.filter_by(id=item_id).one()
+    form = UpdateContactForm(
+        name=contact_item.name,
+        number=contact_item.number,
+        address=contact_item.address,
+        note=contact_item.note
+    )
+
+    if form.validate_on_submit():
+        contact_item.name=form.name.data
+        contact_item.number=form.number.data
+        contact_item.address=form.address.data
+        contact_item.note=form.note.data
+        try:
+            db.session.commit()
+            flash("update successful")
+            return redirect(url_for('contact.index'))
+        except Exception as e:
+            ic(e)
+            flash("error in update")
+            return redirect(url_for('contact.index'))
+    return render_template('contact/update.html', form=form)
